myApp/filters.py
import django_filters
from .models import Artwork, Portrait

class ArtworkFilter(django_filters.FilterSet):
    title = django_filters.CharFilter(lookup_expr='icontains')
    description = django_filters.CharFilter(lookup_expr='icontains')

    class Meta:
        model = Artwork
        fields = ['title', 'description']


class PortraitFilter(django_filters.FilterSet):
    title = django_filters.CharFilter(lookup_expr='icontains')
    description = django_filters.CharFilter(lookup_expr='icontains')

    class Meta:
        model = Portrait
        fields = ['title', 'description']
